+++ pkgs/hooks/test/link_input_test.dart
@@ -34,6 +34,7 @@ void main() async {
       'config': {
         'build_asset_types': ['asset-type-1', 'asset-type-2'],
       },
+      'assets_from_linking': [],
       'out_dir_shared': outputDirectoryShared.toFilePath(),
       'out_file': outFile.toFilePath(),
       'package_name': packageName,
@@ -50,7 +51,11 @@ void main() async {
         outputDirectoryShared: outputDirectoryShared,
       )
       ..config.addBuildAssetTypes(['asset-type-1', 'asset-type-2'])
-      ..setupLink(assets: assets, recordedUsesFile: null);
+      ..setupLink(
+        assetsFromBuilding: assets,
+        recordedUsesFile: null,
+        assetsFromLinking: [],
+      );
     final input = inputBuilder.build();
 
     expect(input.json, inputJson);