+++ pkgs/hooks_runner/lib/src/build_runner/build_runner.dart
@@ -152,11 +152,11 @@ class NativeAssetsBuildRunner {
     /// Key is packageName.
     final globalAssetsForBuild = <String, List<EncodedAsset>>{};
     for (final package in buildPlan) {
-      final assetsForBuild = _assetsForBuildForPackage(
-        packageGraph: packageGraph!,
-        packageName: package.name,
-        globalAssetsForBuild: globalAssetsForBuild,
-      );
+      final dependencies = packageGraph!.neighborsOf(package.name).toSet();
+      final assetsForBuild = <String, List<EncodedAsset>>{
+        for (final entry in globalAssetsForBuild.entries)
+          if (dependencies.contains(entry.key)) entry.key: entry.value,
+      };
 
       final inputBuilder = BuildInputBuilder();
 
@@ -248,6 +248,17 @@ class NativeAssetsBuildRunner {
     Uri? resourceIdentifiers,
     required BuildResult buildResult,
   }) async => _timeAsync('BuildRunner.link', () async {
+    final planResult = await _makePlan(
+      hook: Hook.link,
+      buildResult: buildResult,
+    );
+    if (planResult.isFailure) return planResult.asFailure;
+    final (buildPlan, packageGraph) = planResult.success;
+    if (buildPlan.isEmpty) {
+      // Return eagerly if there are no link hooks at all.
+      return Success(HookResult());
+    }
+
     final loadedUserDefines = await _loadedUserDefines;
     final hookResultUserDefines = await _checkUserDefines(loadedUserDefines);
     if (hookResultUserDefines.isFailure) {
@@ -255,14 +266,19 @@ class NativeAssetsBuildRunner {
     }
     var linkResult = hookResultUserDefines.success;
 
-    final planResult = await _makePlan(
-      hook: Hook.link,
-      buildResult: buildResult,
-    );
-    if (planResult.isFailure) return planResult.asFailure;
-    final (buildPlan, packageGraph) = planResult.success;
-
+    /// The key is the package name of the destination package.
+    final globalAssetsForLink = <String, Map<String, List<EncodedAsset>>>{};
     for (final package in buildPlan) {
+      final dependencies = packageGraph!
+          .inverseNeighborsOf(package.name)
+          .toSet();
+
+      final assetsFromLinking = (globalAssetsForLink[package.name] ?? {})
+          .entries
+          .where((entry) => dependencies.contains(entry.key))
+          .expand((e) => e.value)
+          .toList();
+
       final inputBuilder = LinkInputBuilder();
       for (final e in extensions) {
         e.setupLinkInput(inputBuilder);
@@ -288,9 +304,12 @@ class NativeAssetsBuildRunner {
         outputDirectoryShared: outDirSharedUri,
         userDefines: loadedUserDefines?[package.name],
       );
+
       inputBuilder.setupLink(
-        assets: buildResult.encodedAssetsForLinking[package.name] ?? [],
+        assetsFromBuilding:
+            buildResult.encodedAssetsForLinking[package.name] ?? [],
         recordedUsesFile: resourcesFile?.uri,
+        assetsFromLinking: assetsFromLinking,
       );
 
       final input = inputBuilder.build();
@@ -299,7 +318,6 @@ class NativeAssetsBuildRunner {
         for (final e in extensions) ...await e.validateLinkInput(input),
       ];
       if (errors.isNotEmpty) {
-        print(input.assets.encodedAssets);
         _printErrors('Link input for ${package.name} contains errors', errors);
         return const Failure(HooksRunnerFailure.internal);
       }
@@ -318,9 +336,23 @@ class NativeAssetsBuildRunner {
         buildDirUri,
         outDirUri,
       );
-      if (result.isFailure) return result.asFailure;
+      if (result.isFailure) {
+        return result.asFailure;
+      }
       final (hookOutput, hookDeps) = result.success;
       linkResult = linkResult.copyAdd(hookOutput, hookDeps);
+
+      // Merge the assets into the global asset map.
+      (hookOutput as LinkOutput).assets.encodedAssetsForLink.forEach(
+        (packageName, assetsForPackage) => globalAssetsForLink.update(
+          packageName,
+          (current) => {
+            ...current,
+            ...{package.name: assetsForPackage},
+          },
+          ifAbsent: () => {package.name: assetsForPackage},
+        ),
+      );
     }
 
     final errors = [
@@ -784,23 +816,6 @@ ${compileResult.stdout}
     },
   );
 
-  /// Returns only the assets output as assetForBuild by the packages that are
-  /// the direct dependencies of [packageName].
-  Map<String, List<EncodedAsset>>? _assetsForBuildForPackage({
-    required PackageGraph packageGraph,
-    required String packageName,
-    Map<String, List<EncodedAsset>>? globalAssetsForBuild,
-  }) {
-    if (globalAssetsForBuild == null) {
-      return null;
-    }
-    final dependencies = packageGraph.neighborsOf(packageName).toSet();
-    return {
-      for (final entry in globalAssetsForBuild.entries)
-        if (dependencies.contains(entry.key)) entry.key: entry.value,
-    };
-  }
-
   Future<ValidationErrors> _validate(
     HookInput input,
     HookOutput output,
@@ -850,42 +865,25 @@ ${compileResult.stdout}
   Future<
     Result<(BuildPlan plan, PackageGraph? dependencyGraph), HooksRunnerFailure>
   >
-  _makePlan({required Hook hook, BuildResult? buildResult}) async => _timeAsync(
-    '_makePlan',
-    () async {
-      switch (hook) {
-        case Hook.build:
-          final planner = await _planner;
-          final planResult = await planner.makeBuildHookPlan();
-          if (planResult.isFailure) {
-            return planResult.asFailure;
-          }
-          return Success((planResult.success, planner.packageGraph));
-        case Hook.link:
-          // Link hooks are not run in any particular order.
-          // Link hooks are skipped if no assets for linking are provided.
-          final buildPlan = <Package>[];
-          final skipped = <String>[];
-          final encodedAssetsForLinking = buildResult!.encodedAssetsForLinking;
-          final planner = await _planner;
-          final packagesWithHook = await planner.packagesWithHook(Hook.link);
-          for (final package in packagesWithHook) {
-            if (encodedAssetsForLinking[package.name]?.isNotEmpty ?? false) {
-              buildPlan.add(package);
-            } else {
-              skipped.add(package.name);
+  _makePlan({required Hook hook, BuildResult? buildResult}) async =>
+      _timeAsync('_makePlan', () async {
+        switch (hook) {
+          case Hook.build:
+            final planner = await _planner;
+            final planResult = await planner.makeBuildHookPlan();
+            if (planResult.isFailure) {
+              return planResult.asFailure;
             }
-          }
-          if (skipped.isNotEmpty) {
-            logger.info(
-              'Skipping link hooks from ${skipped.join(', ')}'
-              ' due to no assets provided to link for these link hooks.',
-            );
-          }
-          return Success((buildPlan, null));
-      }
-    },
-  );
+            return Success((planResult.success, planner.packageGraph));
+          case Hook.link:
+            final planner = await _planner;
+            final planResult = await planner.makeLinkHookPlan();
+            if (planResult.isFailure) {
+              return planResult.asFailure;
+            }
+            return Success((planResult.success, planner.packageGraph));
+        }
+      });
 
   Future<Result<HookOutput, HooksRunnerFailure>> _readHookOutputFromUri(
     Hook hook,