+++ pkgs/hooks_runner/test/build_runner/build_runner_cycle_test.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:hooks_runner/src/model/hook_result.dart' show HookResult;
 import 'package:logging/logging.dart';
 import 'package:test/test.dart';
 
@@ -11,7 +12,7 @@ import 'helpers.dart';
 const Timeout longTimeout = Timeout(Duration(minutes: 5));
 
 void main() async {
-  test('cycle', timeout: longTimeout, () async {
+  test('build cycle', timeout: longTimeout, () async {
     await inTempDir((tempUri) async {
       await copyTestProjects(targetUri: tempUri);
       final packageUri = tempUri.resolve('cyclic_package_1/');
@@ -33,6 +34,35 @@ void main() async {
           contains(
             'Cyclic dependency for native asset builds in the following '
             'packages: [cyclic_package_1, cyclic_package_2]',
+          ),
+        );
+      }
+    });
+  });
+
+  test('link cycle', timeout: longTimeout, () async {
+    await inTempDir((tempUri) async {
+      await copyTestProjects(targetUri: tempUri);
+      final packageUri = tempUri.resolve('cyclic_link_package_1/');
+
+      await runPubGet(workingDirectory: packageUri, logger: logger);
+
+      {
+        final logMessages = <String>[];
+        final result = await link(
+          buildResult: HookResult(),
+          packageUri,
+          createCapturingLogger(logMessages, level: Level.SEVERE),
+          dartExecutable,
+          buildAssetTypes: [],
+        );
+        final fullLog = logMessages.join('\n');
+        expect(result.isFailure, isTrue);
+        expect(
+          fullLog,
+          contains(
+            'Cyclic dependency for native asset links in the following '
+            'packages: [cyclic_link_package_1, cyclic_link_package_2]',
           ),
         );
       }