+++ pkgs/hooks_runner/test_data/link_inverse_package/hook/link.dart
@@ -1,0 +1,11 @@
+// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:hooks/hooks.dart';
+
+void main(List<String> arguments) async {
+  await link(arguments, (input, output) async {
+    output.metadata.add('link_inverse_app', 'key', 'value');
+  });
+}