+++ pkgs/code_assets/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.19.5-wip
+## 0.20.0-wip
 
 - Bump `package:hooks` to 0.20.0.